ベルの動作と取得効果を一部追加

+++ to_win_bee/EChara.cpp
@@ -7,19 +7,26 @@
 
 //-------------------------------------------------------------------
 //プレイヤーとの当たり判定
-void EChara::HitPlayer()
+//引数：プレイヤーが死ぬか否か(bool)
+//戻り値：プレイヤーに当たったか否か
+bool EChara::HitPlayer(bool killPlayer)
 {
 	auto player = ge->GetTask_One_GN<Player::Object>("本編", "プレイヤー");
 	if (player == nullptr)
-		return;
+		return false;
 
 	ML::Box2D eh = hitBase.OffsetCopy(pos);
 	ML::Box2D ph = player->hitBase.OffsetCopy(player->pos);
 
 	if (eh.Hit(ph))
 	{
-		//player->state = Death;
-		//player->cntTime = 0;
-		player->Kill(); //仮
+		if (killPlayer)
+		{
+			//player->state = Death;
+			//player->cntTime = 0;
+			player->Kill(); //仮
+		}
+		return true;
 	}
+	return false;
 }
